refactor cam dir
