+++ src/components/cam/camSelect.tsx
@@ -1,0 +1,47 @@
+import { useContext } from "react";
+import { CamContext } from "./context";
+
+export const CamSelect = () => {
+  const { devices, videoRef, flipRef } = useContext(CamContext);
+
+  return (
+    <>
+      <label>Select Camera</label>
+      <select
+        className="form-select"
+        onChange={(e) => {
+          navigator.mediaDevices
+            .getUserMedia({
+              video: {
+                deviceId: e.currentTarget.value,
+              },
+            })
+            .then((stream) => {
+              videoRef.current.srcObject = stream;
+            });
+        }}
+      >
+        {devices.map((d, idx) => (
+          <option key={d.value} value={d.value}>
+            {idx + 1}. {d.label || "Camera"}
+          </option>
+        ))}
+      </select>
+      <div className="form-check">
+        <label className="form-check-label">
+          <input
+            className="form-check-input"
+            type="checkbox"
+            onChange={(e) => {
+              if (e.currentTarget.checked)
+                videoRef.current.style.transform = "scaleX(-1)";
+              else videoRef.current.style.transform = "";
+              flipRef.current = e.currentTarget.checked;
+            }}
+          />
+          Horizontal Flip
+        </label>
+      </div>
+    </>
+  );
+};
