+++ src/components/cam/type.ts
@@ -1,0 +1,4 @@
+export type CamDataHandler = (input: HTMLVideoElement) => Promise<void>;
+export type SetCamDataHandler = (cameraHandler: CamDataHandler) => void;
+export type Clear = () => void;
+
